changement pour utilisation des factory et des repositories

+++ src/main/java/ulaval/glo2003/domain/product/Amount.java
@@ -24,6 +24,6 @@ public class Amount {
     }
 
     public double getValue() {
-        return value;
+        return value.doubleValue();
     }
 }
